+++ crates/twirp/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod client;
 pub mod error;
 pub mod headers;
+pub mod mock;
 pub mod server;
 
 #[cfg(any(test, feature = "test-support"))]
